backend/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
Pillow==10.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
dj-database-url==2.1.0